+++ test/plug/router_test.exs
@@ -105,10 +105,6 @@ defmodule Plug.RouterTest do
       resp(conn, 200, inspect(bar))
     end
 
-    get "/5/bar-*bar" do
-      resp(conn, 200, inspect(bar))
-    end
-
     match "/6/bar" do
       resp(conn, 200, "ok")
     end
@@ -119,22 +115,6 @@ defmodule Plug.RouterTest do
 
     get "/8/bar baz/:bat" do
       resp(conn, 200, bat)
-    end
-
-    get "/9/:bar.json" when bar in ["foo", "bar"] or bar == "baz" do
-      if bar == "baz" do
-        resp(conn, 200, inspect(bar <> " from or boolean guard"))
-      else
-        resp(conn, 200, inspect(bar <> " from in guard"))
-      end
-    end
-
-    get "/9/:bar.json" when bar == "specific_value" and bar not in ["other_value"] do
-      resp(conn, 200, inspect(bar <> " from == guard"))
-    end
-
-    get "/9/:bar.json" when bar != "value" do
-      resp(conn, 200, inspect(bar <> " from != guard"))
     end
 
     get "/9/:bar.json" do
@@ -155,15 +135,6 @@ defmodule Plug.RouterTest do
 
     get "/11/:bar.js.map" do
       resp(conn, 200, inspect(bar))
-    end
-
-    get "/11/:foo/:bar@app/baz/:bat.js.map"
-        when foo in ["foo"] and bar == "bar" and bat == "bat" do
-      resp(
-        conn,
-        200,
-        ([inspect(foo), inspect(bar), inspect(bat)] |> Enum.join(" ")) <> " from boolean guards"
-      )
     end
 
     get "/11/:foo/:bar@app/baz/:bat.js.map" do
@@ -323,46 +294,12 @@ defmodule Plug.RouterTest do
     assert conn.resp_body == "oops"
   end
 
-  test "dispatch dynamic segment with suffix and var guard" do
-    conn = call(Sample, conn(:get, "/9/other_value.json"))
-    assert conn.resp_body == ~s("other_value from != guard")
-    assert conn.params == %{"bar" => "other_value"}
-    assert conn.path_params == %{"bar" => "other_value"}
-
-    conn = call(Sample, conn(:get, "/9/specific_value.json"))
-    assert conn.resp_body == ~s("specific_value from == guard")
-
-    conn = call(Sample, conn(:get, "/9/foo.json"))
-    assert conn.resp_body == ~s("foo from in guard")
-
-    conn = call(Sample, conn(:get, "/9/bar.json"))
-    assert conn.resp_body == ~s("bar from in guard")
-
-    conn = call(Sample, conn(:get, "/9/baz.json"))
-    assert conn.resp_body == ~s("baz from or boolean guard")
-  end
-
-  test "dispatch dynamic segment with/without suffix and guards" do
-    conn = call(Sample, conn(:get, "/11/foo/bar@app/baz/bat.js.map"))
-    assert conn.resp_body == ~s("foo\" "bar" "bat" from boolean guards)
-    assert conn.params == %{"bar" => "bar", "bat" => "bat", "foo" => "foo"}
-    assert conn.path_params == %{"bar" => "bar", "bat" => "bat", "foo" => "foo"}
-
-    conn = call(Sample, conn(:get, "/11/boo/boo-app/baz/boo.js.map"))
-    assert conn.resp_body == "oops"
-  end
-
   test "dispatch glob segment" do
     conn = call(Sample, conn(:get, "/4/value"))
     assert conn.resp_body == ~s(["value"])
 
     conn = call(Sample, conn(:get, "/4/value/extra"))
     assert conn.resp_body == ~s(["value", "extra"])
-  end
-
-  test "dispatch glob segment with prefix" do
-    conn = call(Sample, conn(:get, "/5/bar-value/extra"))
-    assert conn.resp_body == ~s(["bar-value", "extra"])
   end
 
   test "dispatch custom route" do